tests/wpt/tests/html/canvas/element/text/2d.text.measure.lang.inherit.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.measure.lang.inherit</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<style>
@font-face {
  font-family: Lato-Medium;
  src: url("/fonts/Lato-Medium.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.lang.inherit</h1>
<p class="desc">Testing the lang attribute</p>


<span style="font-family: Lato-Medium; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"lang="tr"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
promise_test(async t => {

  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  await document.fonts.ready;

  ctx.font = '50px Lato-Medium';
  ctx.lang = 'inherit';
  const text = 'fi';
  const tm_tr = ctx.measureText(text);
  const tr_width = tm_tr.width;

  ctx.lang = 'en';
  const tm_en = ctx.measureText(text);
  const en_width = tm_en.width;

  _assert(tr_width > en_width, "tr_width > en_width");

}, "Testing the lang attribute");
</script>
